// angular-front/ptx-front/src/app/shared/models/weather-class.enum.ts
export enum WeatherClassEnum {
    sunny = 'we-sunny',
    mostlysunny = 'we-mostly-sunny',
    partlysunny = 'we-partly-sunny',
    cloudy = 'we-cloudy',
    fog = 'we-fog',
    showersday = 'we-showers-day',
    sunnyshowers = 'we-sunny-showers',
    tstormday = 'we-t-storm-day',
    tstormnight = 'we-t-storm-night',
    rainday = 'we-rain-day',
    rainnight = 'we-rain-night',
    flurriesday = 'we-flurries-day',
    flurriesnight = 'we-flurries-night',
    snow = 'we-snow',
    freezingrain = 'we-freezing-rain',
    windy = 'we-windy',
    clear = 'we-clear',
    mostlyclear = 'we-mostly-clear',
    partlyclear = 'we-partly-clear',
    showersnight = 'we-showers-night'
}